requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
websockets>=12.0
yt-dlp>=2023.11.16
boto3>=1.29.7
python-multipart>=0.0.6
pydantic>=2.5.0
pydantic-settings>=2.1.0
aiofiles>=23.2.1
httpx>=0.25.0
imageio-ffmpeg>=0.4.9
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9
pymysql>=1.1.0
python-dotenv>=1.0.0
supabase>=2.0.0
PyJWT>=2.8.0


